+++ src/lte/model/lte-miesm-error-model.h
@@ -78,7 +78,7 @@ namespace ns3 {
          * \param miHistory  MI of past transmissions (in case of retx)
          * \return the TB error rate and MI
          */
-        static TbStats_t GetTbDecodificationStats (const SpectrumValue& sinr, const std::vector<int>& map, uint16_t prb_size, uint16_t size, uint8_t mcs, HarqProcessInfoList_t miHistory, uint8_t num, std::string chan, double speed);
+        static TbStats_t GetTbDecodificationStats (const SpectrumValue& sinr, const std::vector<int>& map, uint16_t prb_size, uint32_t size, uint8_t mcs, HarqProcessInfoList_t miHistory, uint8_t num, std::string chan, double speed);
         /**
         * \brief load error-model data from the input json file
         *
